+++ R/ast2df.R
@@ -1,28 +1,53 @@
-# Implemented but not used yet
+#' Flattens attribute
+#'
+#' Implemented but not used yet
+#'
+#' @param a A set of ttributes. Typically `x$c[[1]]`
+#'
+#' @noRd
 flatten_attr <- function(a) {
   c(
     list(id = a[[1]], class = unlist(a[[2]])),
-    stats::setNames(map_chr(a[[3]], 2), map_chr(a[[3]], 1))
+    stats::setNames(purrr::map_chr(a[[3]], 2), purrr::map_chr(a[[3]], 1))
   )
 }
 
-add_type <- function(x, t) {
-  has_link = x$t %in% c('Image', 'Link')
+is_branch <- function(x) {
+  nm = names(x)
+  identical(nm, c('t', 'c')) || identical(nm, 't')
+}
 
-  x$t <- c(
-    structure(list(if (has_link) x$c[[3]][[1]] else TRUE), .Names = x$t),
-    if (is.list(t)) t else structure(list(TRUE), .Names = t)
+has_attr <- function(x) {
+  !(is.atomic(x) || is_branch(x) || all(vapply(x, is_branch, NA)))
+}
+
+add_type <- function(x, t) {
+  parents <- if (is.list(t)) t else stats::setNames(list(TRUE), t)
+  child <- stats::setNames(
+    list(structure(
+      if (isTRUE(x$t %in% c('Image', 'Link'))) x$c[[3]][[1]] else TRUE,
+      pandoc_attr = if (has_attr(x$c)) {
+        flatten_attr(x$c[[1]])
+      } else {
+        NULL
+      }
+    )),
+    x$t
   )
 
-  if (has_link) {
-    x$c <- x$c[[2]]
-  }
+  x$t <- c(child, parents)
+
+  if (has_attr(x$c)) x$c <- x$c[[2]]
 
   x
 }
 
 resolve_type <- function(x) {
   if (is.atomic(x$c)) return(x)
+  if (identical(x$c, list())) {
+    x$c <- ''
+    return(x)
+  }
   if (identical(names(x$c), c('t', 'c'))) return(add_type(x$c, x$t))
   return(lapply(x$c, add_type, x$t))
 }