+++ R/parse-md.R
@@ -11,6 +11,12 @@ vertical_align <- function(sup, sub) {
   )
 }
 
+pandoc_attr <- function(x, y) {
+  a = attr(x, 'pandoc_attr', exact = TRUE)
+  if (is.null(a)) return(NULL)
+  a[[y]]
+}
+
 parse_md_ <- function(x, .from = 'markdown', auto_color_link = 'blue') {
   if (!is.character(auto_color_link) || length(auto_color_link) != 1) {
     stop('`auto_color_link` must be a string')
@@ -25,12 +31,15 @@ parse_md_ <- function(x, .from = 'markdown', auto_color_link = 'blue') {
       italic = y$Emph %||% NA,
       bold = y$Strong %||% NA,
       url = y$Link %||% NA_character_,
+      width = as.numeric(pandoc_attr(y$Image, 'width') %||% NA_real_),
+      height = as.numeric(pandoc_attr(y$Image, 'height') %||% NA_real_),
       vertical.align = vertical_align(y$Superscript, y$Subscript),
       stringsAsFactors = FALSE
     )
   ) %>%
     dplyr::mutate(
-      color = dplyr::if_else(is.na(url), NA_character_, auto_color_link)
+      color = dplyr::if_else(is.na(url), NA_character_, auto_color_link),
+      img_data = y$Image %||% list(list())
     )
 }