+++ R/parse-md.R
@@ -11,7 +11,17 @@ vertical_align <- function(sup, sub) {
   )
 }
 
-parse_md_ <- function(x, .from = 'markdown') {
+pandoc_attr <- function(x, y) {
+  a = attr(x, 'pandoc_attr', exact = TRUE)
+  if (is.null(a)) return(NULL)
+  a[[y]]
+}
+
+parse_md_ <- function(x, .from = 'markdown', auto_color_link = 'blue') {
+  if (!is.character(auto_color_link) || length(auto_color_link) != 1) {
+    stop('`auto_color_link` must be a string')
+  }
+
   y <- x %>% md2ast(.from = .from) %>% ast2df %>% as.list
 
   dplyr::bind_rows(
@@ -20,12 +30,22 @@ parse_md_ <- function(x, .from = 'markdown') {
       txt = y$txt,
       italic = y$Emph %||% NA,
       bold = y$Strong %||% NA,
+      url = y$Link %||% NA_character_,
+      width = as.numeric(pandoc_attr(y$Image, 'width') %||% NA_real_),
+      height = as.numeric(pandoc_attr(y$Image, 'height') %||% NA_real_),
       vertical.align = vertical_align(y$Superscript, y$Subscript),
       stringsAsFactors = FALSE
     )
-  )
+  ) %>%
+    dplyr::mutate(
+      color = dplyr::if_else(is.na(url), NA_character_, auto_color_link),
+      img_data = y$Image %||% list(list())
+    )
 }
 
-parse_md <- function(x, .from = 'markdown') {
-  structure(lapply(x, parse_md_, .from = .from), class = 'paragraph')
+parse_md <- function(x, .from = 'markdown', auto_color_link = 'blue') {
+  structure(
+    lapply(x, parse_md_, .from = .from, auto_color_link = auto_color_link),
+    class = 'paragraph'
+  )
 }